Cuenta los sindicatos activos y los afiliados en cada mes en un periodo de años

Cuenta los sindicatos activos y los afiliados en cada mes en un periodo de años

activos_historico(desde = 2020, hasta = 2021)

Arguments

desde

Año desde el cuál partir haciendo el conteo. 2020 por defecto

hasta

Año hasta el cuál partir haciendo el conteo. 2021 por defecto

Value

Una tibble con el número de sindicatos y socios en cada mes del periodo indicado

Examples

activos_historico(desde=2020,hasta=2021)
#>     ano mes sindicatos afiliads
#> 1  2020   1      12589  1265533
#> 2  2020   2      12550  1263785
#> 3  2020   3      12535  1264111
#> 4  2020   4      12485  1261895
#> 5  2020   5      12485  1261851
#> 6  2020   6      12499  1262408
#> 7  2020   7      12505  1262893
#> 8  2020   8      12509  1262790
#> 9  2020   9      12520  1263401
#> 10 2020  10      12547  1264482
#> 11 2020  11      12587  1265836
#> 12 2020  12      12614  1266958
#> 13 2021   1      12628  1267914
#> 14 2021   2      12624  1268030
#> 15 2021   3      12635  1268480
#> 16 2021   4      12638  1269175
#> 17 2021   5      12658  1270999
#> 18 2021   6      12675  1271929
#> 19 2021   7      12715  1273070
#> 20 2021   8      12770  1275135
#> 21 2021   9      12813  1275582
#> 22 2021  10      12857  1277331
#> 23 2021  11      12628  1264765
#> 24 2021  12      12373  1247580