Agrupa por empresa e indica N de sindicatos activos y afiliados/as

Además reporta para cada empresa el tamaño promedio de los sindicatos.

empresas_sindicatos_activos()

Value

Una tibble en la que cada fila es una empresa

Examples

empresas_sindicatos_activos()
#> # A tibble: 4,571 x 4
#>    rut_empresa sindicatos afiliads tamano_promedio
#>    <chr>            <int>    <dbl>           <dbl>
#>  1 81537600            73     9626           132. 
#>  2 76134941            49     9212           188  
#>  3 17005635            46     6426           140. 
#>  4 76879810            33     3942           119. 
#>  5 61704000            31    13808           445. 
#>  6 83150900            29      839            28.9
#>  7 96973670            20     1380            69  
#>  8 78627210            19     1966           103. 
#>  9 81496800            19     1352            71.2
#> 10 96843130            18     2027           113. 
#> # ... with 4,561 more rows