Agrega información del Servicio de impuestos internos

Se filtra la data a los sindicatos activos y de empresa o establecimiento. Luego se agregan variables del Servicio de Impuestos Internos relacionadas a ventas, número de trabajadores y actividad económica.

sii_agregar()

Value

Una data frame donde las filas son los sindicatos. Se agrega info de sector económico de empresa, ventas, número de trabajadores y tasa de sindicalización del sindicato en la empresa

Examples

datos <- sii_agregar()
table(datos$rubro_sii,useNA="ifany")
#> 
#>                                                                                            
#>                                                                                          0 
#>                                   ACTIVIDADES ARTISTICAS, DE ENTRETENIMIENTO Y RECREATIVAS 
#>                                                                                         98 
#>                                        ACTIVIDADES DE ALOJAMIENTO Y DE SERVICIO DE COMIDAS 
#>                                                                                        203 
#>                          ACTIVIDADES DE ATENCION DE LA SALUD HUMANA Y DE ASISTENCIA SOCIAL 
#>                                                                                        212 
#>   ACTIVIDADES DE LOS HOGARES COMO EMPLEADORES; ACTIVIDADES NO DIFERENCIADAS DE LOS HOGARES 
#>                                                                                          0 
#>                                 ACTIVIDADES DE ORGANIZACIONES Y ORGANOS EXTRATERRITORIALES 
#>                                                                                          3 
#>                                        ACTIVIDADES DE SERVICIOS ADMINISTRATIVOS Y DE APOYO 
#>                                                                                        667 
#>                                                       ACTIVIDADES FINANCIERAS Y DE SEGUROS 
#>                                                                                        161 
#>                                                                  ACTIVIDADES INMOBILIARIAS 
#>                                                                                         89 
#>                                          ACTIVIDADES PROFESIONALES, CIENTIFICAS Y TECNICAS 
#>                                                                                        239 
#>     ADMINISTRACION PUBLICA Y DEFENSA; PLANES DE SEGURIDAD SOCIAL DE AFILIACION OBLIGATORIA 
#>                                                                                          2 
#>                                               AGRICULTURA, GANADERIA, SILVICULTURA Y PESCA 
#>                                                                                        333 
#>   COMERCIO AL POR MAYOR Y AL POR MENOR; REPARACION DE VEHICULOS AUTOMOTORES Y MOTOCICLETAS 
#>                                                                                        392 
#>                                                                               CONSTRUCCION 
#>                                                                                        248 
#>                                                                                  ENSEÑANZA 
#>                                                                                        942 
#>                                                            EXPLOTACION DE MINAS Y CANTERAS 
#>                                                                                        233 
#>                                                                    INDUSTRIA MANUFACTURERA 
#>                                                                                       1375 
#>                                                               INFORMACION Y COMUNICACIONES 
#>                                                                                        148 
#>                                                             OTRAS ACTIVIDADES DE SERVICIOS 
#>                                                                                        370 
#> SUMINISTRO DE AGUA; EVACUACION DE AGUAS RESIDUALES, GESTION DE DESECHOS Y DESCONTAMINACION 
#>                                                                                         49 
#>                                SUMINISTRO DE ELECTRICIDAD, GAS, VAPOR Y AIRE ACONDICIONADO 
#>                                                                                        105 
#>                                                                TRANSPORTE Y ALMACENAMIENTO 
#>                                                                                        714 
#>                                                                          Valor por Defecto 
#>                                                                                         13 
#>           COMERCIO AL POR MAYOR Y AL POR MENOR; REPARACION DE VEHICULOS AUTOMOTORES Y MOTO 
#>                                                                                        436 
#>           SUMINISTRO DE AGUA; EVACUACION DE AGUAS RESIDUALES, GESTION DE DESECHOS Y DESCON 
#>                                                                                         75 
#>           ADMINISTRACION PUBLICA Y DEFENSA; PLANES DE SEGURIDAD SOCIAL DE AFILIACION OBLIG 
#>                                                                                         14 
#>           ACTIVIDADES DE LOS HOGARES COMO EMPLEADORES; ACTIVIDADES NO DIFERENCIADAS DE LOS 
#>                                                                                          0 
#>                                                                                       <NA> 
#>                                                                                        281 
table(datos$ventas_sii,useNA="ifany")
#> 
#>    1    2    3    4    5    6    7    8    9   10   11   12   13 <NA> 
#>  583   65   48   98   79  153  254  392  521  710 1184  573 2461  281 
nrow(datos[datos$tasa_sindicalizacion<=1&!is.na(datos$tasa_sindicalizacion),])
#> [1] 6657
summary(datos[datos$tasa_sindicalizacion<=1,]$tasa_sindicalizacion)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#>  0.0000  0.0530  0.1630  0.2311  0.3595  1.0000     598